5.4 random items update, num_boss_adds can now be a float value

+++ public/javascripts/core.js
@@ -1436,7 +1436,7 @@
         num_boss_adds: {
           name: "Number of Boss Adds",
           desc: "Used for the Trinket Sigil of Rampage and the Legendary Cloak",
-          datatype: 'integer',
+          datatype: 'float',
           type: 'input',
           min: 0,
           max: 5,
